+++ CHANGELOG.md
@@ -40,6 +40,7 @@ The following configuration changes in the `[data]` section may need to changed 
 - [#7320](https://github.com/influxdata/influxdb/issues/7320): Update defaults in config for latest best practices
 - [#7495](https://github.com/influxdata/influxdb/pull/7495): Rewrite regexes of the form host = /^server-a$/ to host = 'server-a', to take advantage of the tsdb index.
 - [#6704](https://github.com/influxdata/influxdb/issues/6704): Optimize first/last when no group by interval is present.
+- [#4461](https://github.com/influxdata/influxdb/issues/4461): Change default time boundaries for raw queries.
 
 ### Bugfixes